+++ rich/__main__.py
@@ -90,7 +90,9 @@ def make_test_card() -> Table:
 
     table.add_row(
         "CJK support",
-        Panel("该库支持中文，日文和韩文文本！", expand=False, style="red", box=box.DOUBLE_EDGE,),
+        Panel(
+            "该库支持中文，日文和韩文文本！", expand=False, border_style="red", box=box.DOUBLE_EDGE,
+        ),
     )
 
     emoji_example = (