+++ tests/test_console.py
@@ -172,18 +172,18 @@ def test_justify_renderable_none():
 
 def test_justify_renderable_left():
     console = Console(
-        file=io.StringIO(), force_terminal=True, width=20, legacy_windows=False
+        file=io.StringIO(), force_terminal=True, width=10, legacy_windows=False
     )
     console.print(Panel("FOO", expand=False), justify="left")
-    assert console.file.getvalue() == "╭───╮\n│FOO│\n╰───╯\n"
+    assert console.file.getvalue() == "╭───╮     \n│FOO│     \n╰───╯     \n"
 
 
 def test_justify_renderable_center():
     console = Console(
-        file=io.StringIO(), force_terminal=True, width=20, legacy_windows=False
+        file=io.StringIO(), force_terminal=True, width=10, legacy_windows=False
     )
     console.print(Panel("FOO", expand=False), justify="center")
-    assert console.file.getvalue() == "       ╭───╮\n       │FOO│\n       ╰───╯\n"
+    assert console.file.getvalue() == "  ╭───╮   \n  │FOO│   \n  ╰───╯   \n"
 
 
 def test_justify_renderable_right():